requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
decorator==5.1.1
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.2
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
django-allauth
